+++ versioned/storage/dynamodb/src/main/java/org/projectnessie/versioned/storage/dynamodb/DynamoDBConstants.java
@@ -42,8 +42,5 @@ final class DynamoDBConstants {
   static final String CONDITION_STORE_REF = "attribute_not_exists(" + COL_REFERENCES_POINTER + ")";
   static final String CONDITION_STORE_OBJ = "attribute_not_exists(" + COL_OBJ_TYPE + ")";
 
-  static final String CONDITIONAL_UPDATE_DELETE_CONDITION =
-      "(" + COL_OBJ_TYPE + " = :objtype) AND (" + COL_OBJ_VERS + " = :objvers)";
-
   private DynamoDBConstants() {}
 }